requirements.txt
requests>=2.25.1
Pillow>=9.0.0
streamlit>=1.25.0
firebase-admin>=6.3.0
opencv-python>=4.5.5.64
python-dotenv>=1.0.0
